requirements.txt
google-generativeai          0.8.3
Flask                        3.1.0
python-dotenv                1.0.1
gunicorn                     23.0.0